% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{assert_config_file_appropriate}
\alias{assert_config_file_appropriate}
\title{Check config filepath is appropriate}
\usage{
assert_config_file_appropriate(file)
}
\arguments{
\item{file}{path to config file (string)}
}
\value{
invisibly returns TRUE
}
\description{
Ensures file is hidden, in a folder that exists, and that has read and write permissions.
The file itself doesn't already have to exist
}
