% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{config_filepath}
\alias{config_filepath}
\title{Default database configuration yaml}
\usage{
config_filepath()
}
\value{
path to default config banner (string)
}
\description{
Default database configuration yaml
}
\examples{
config_filepath()
}
