% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_add_table}
\alias{db_add_table}
\title{Add a table and write to a database.}
\usage{
db_add_table(tbl_name, tbl_path, src, overwrite = !append,
  append = !overwrite, ...)
}
\arguments{
\item{tbl_name}{String. The name of the table.}

\item{tbl_path}{String. The path to the table.}

\item{src}{A database source.}

\item{overwrite}{Flag. Should an existing table be overwritten? Defaults to
\code{!append}}

\item{append}{Flag. Should an existing table be append? Defaults to
\code{!overwrite}.}

\item{...}{Further arguments passed to \link[data.table]{fread}.}
}
\description{
Reads a plain text table and writes it to a database. Future versions will
automate chuncked reads for very large tables. Currently uses
\link[data.table]{fread} for quickly reading plain text tables.
}

