% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_build}
\alias{db_build}
\title{Build a database given a configuration file}
\usage{
db_build(cnf)
}
\arguments{
\item{cnf}{String. A path to easydb configuration file.}
}
\description{
Builds an SQLite database given an easydb configuration file
}

