% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_config}
\alias{db_config}
\title{Read an EasyDB configuration file}
\usage{
db_config(cnf)
}
\arguments{
\item{cnf}{Path to an EasyDB configuration file, or the result of
\link{db_config}.}
}
\description{
Reads an EasyDB configuration file and returns a 'dbcnf' object.
}

