% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_connect}
\alias{db_connect}
\title{Create a connection to a DBMS}
\usage{
db_connect(drv = c("SQLite", "MariaDB", "MySQL", "Odbc", "Oracle",
  "PostgreSQL"), ...)
}
\arguments{
\item{drv}{a character used to match partially a DBMS name (ignoring case)
and match it to a default driver, OR  \emph{DBIDriver} object (as with \code{DBI::dbConnect}),
OR a list of specs (in which case other parameters are ignored)}

\item{...}{authentication arguments needed by the DBMS instance;
supported parameters depend on driver and are enumerated in \code{?RSQLite::SQLite},
\code{?RMariaDB::MariaDB}, \code{RMySQL::MySQL}, \code{?odbc::odbc}, \code{ROracle::Oracle},
\code{?RPostgres::Postgres} or in the help file of the appropriate driver.}
}
\value{
an S4 object that inherits from DBIConnection. This object is used to
communicate with the database engine.
}
\description{
It's a simple wrapper around \code{DBI::dbConnect} that makes it easy to use specs
stored in a list and makes it easier to choose the relevant driver.
}
\examples{
# using only defaults
db_connect(path=":memory:")
# feeding a list of specs to the drv argument
db_connect(list(path=":memory:"))
# partial matching
db_connect(drv = "sqli", path=":memory:")
# explicit driver
db_connect(drv = RSQLite::SQLite(), path=":memory:")
}
