% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_doctor}
\alias{db_doctor}
\title{Perform checkup on database}
\usage{
db_doctor(cnf)
}
\arguments{
\item{cnf}{Path to an easydb configuration file, or the result of
\link{db_config}.}
}
\description{
Runs all tests and checks for duplicated keys.
}

