% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_dump}
\alias{db_dump}
\title{Dump an SQLite database to CSVs}
\usage{
db_dump(src, dir)
}
\arguments{
\item{src}{A database source.}

\item{dir}{String. Directory to write CSVs.}
}
\description{
Dumps each table in an SQLite database to CSVs
}

