% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_import_table}
\alias{db_import_table}
\title{Import a table from text file to database.}
\usage{
db_import_table(tbl_path, tbl_name, src, overwrite = !append,
  append = !overwrite, ...)
}
\arguments{
\item{tbl_path}{String. The path to the table.}

\item{tbl_name}{String. The name of the table.}

\item{src}{A database source.}

\item{overwrite}{Flag. Should an existing table be overwritten? Defaults to
\code{!append}}

\item{append}{Flag. Should an existing table be append? Defaults to
\code{!overwrite}.}

\item{...}{Further arguments passed to \link[data.table]{fread}.}
}
\description{
Reads a plain text table and writes it to a database. Future versions will
automate chuncked reads for very large tables. Currently uses
\link[data.table]{fread} for quickly reading plain text tables.
}

