% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easydb.R
\name{easydb}
\alias{easydb}
\alias{$.DBIConnection}
\alias{$<-.DBIConnection}
\alias{!.tbl_lazy}
\title{Easy DB workflow using a few compact methods}
\description{
We define \code{$.DBIConnection()}, \code{$<-.DBIConnection()} and
\code{!.tbl_lazy()} to be able to :
\itemize{
\item write in a data base with
\code{con$foo <- my_data_frame} (where \code{con} is a connection built with
\code{DBI::dbConnect()} or this package's db_connect), short for
\code{DBI::dbWriteTable(con, "foo", my_data_frame)}
\item build a lazy table with \code{con$foo}, short for \code{dplyr::tbl(con, "foo")},
so it can be used to do some server side data manipulation using \emph{dplyr}
\item collect a lazy table into the R session with \code{!my_lazy_table} (or directly
\code{!con$foo}), equivalent to \code{dplyr::collect(my_lazy_table)} or
\code{dplyr::collect(dplyr::tbl(con, "foo"))}
}
}
\examples{
if(requireNamespace("RSQLite")){
db <- db_connect("sqlite", path = ":memory:")
db$foo <- head(cars,2)
db$foo
!db$foo
}
}
