% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{easydb_available_databases}
\alias{easydb_available_databases}
\title{Available Databases}
\usage{
easydb_available_databases(config_file = config_filepath())
}
\arguments{
\item{config_file}{path to yaml filecontaining configuration information about databases (port, host, etc. ) (string)}
}
\value{
vector of database names
}
\description{
List databases with configuration details stored in a configuration file.
}
\examples{
easydb_available_databases()
}
