% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{easydb_connect}
\alias{easydb_connect}
\title{Easy Database Connection}
\usage{
easydb_connect(dbname, config_file = config_filepath(), from_scratch = FALSE)
}
\arguments{
\item{dbname}{name of database (string)}

\item{config_file}{path to yaml filecontaining configuration information about databases (port, host, etc. ) (string)}

\item{from_scratch}{should we delete any current config / credentials for databases with the supplied name and start again? (logical)}
}
\value{
connection to database (connection)
}
\description{
Easy Database Connection
}
\examples{
if(interactive()) {
  # Connect to SQLite database
  path_to_db <- system.file(package = 'easydb', 'testdbs/mtcars.sqlite')
  con <- easydb_connect(dbname = path_to_db)

  # Disconnect from database when finished
  easydb_disconnect(con)
}
}
