% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{easydb_default_config_filepath}
\alias{easydb_default_config_filepath}
\title{Default Config File}
\usage{
easydb_default_config_filepath()
}
\value{
invisible returns the path to default config file (string)
}
\description{
Describe full path to default config file
}
\examples{
easydb_default_config_filepath()
}
