% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{easydb_disconnect}
\alias{easydb_disconnect}
\title{Disconnect from database}
\usage{
easydb_disconnect(connection)
}
\arguments{
\item{connection}{coneection generated by [easydb_connect]}
}
\value{
Invisibly returns TRUE
}
\description{
Simple wrapper around [DBI::dbDisconnect()]
}
