% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{easydb_init}
\alias{easydb_init}
\title{Initialise Configuration Store}
\usage{
easydb_init(config_file)
}
\arguments{
\item{config_file}{path to create a new config file to store database configurations}
}
\value{
invisibly returns path to config file
}
\description{
Create a file to store your database configurations in.
}
\examples{

# Choose config file path
# Do NOT use tempfile in practice.
# Choose a fixed location such as '~/.easydb'
config <- tempfile('.example_config')

# Initialise Configuration File
easydb_init(config)

}
