% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql.R
\name{psql}
\alias{psql}
\title{Run a SQL query on a postgres DB and return a nicely formatted tibble.}
\usage{
psql(query, dbname = "mydb", UTC = TRUE, ...)
}
\arguments{
\item{query}{A valid SQL query}

\item{dbname}{The database you wish to query. Connection parameters will be loaded from the pg_service.conf file and used. See \url{http://www.postgresql.org/docs/9.4/static/libpq-pgservice.html} for details on this file and syntax.}

\item{UTC}{A logical value indicating whether the timestamps should be returned in UTC side. Defaults to \code{TRUE}. Currently \code{FALSE} will return Pacific time.}

\item{...}{Other arguments passed on to methods}
}
\value{
\href{https://github.com/hadley/tibble}{tibble} containing the query results
}
\description{
Returns the results in a tibble.
}
\examples{
psql(query = "SELECT * FROM consumers limit 3;", dbname = 'mydb')


}

