% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{query_db}
\alias{query_db}
\title{Create a function to query a specific database}
\usage{
query_db(dbname, UTC = TRUE)
}
\arguments{
\item{dbname}{The database you wish to query}

\item{UTC}{Logical value for whether you want UTC or Pacific timezone.}
}
\value{
function prepped to query the database you want.
}
\description{
The function will return a function that can be used to query a database. The example should clear up any confusion.
}
\details{
Because doing an analysis normally means working on just one database, it is annoying to have to continuously type \code{psql(query, dbname = 'myproduct')}, adding in the dbname 'myproduct' each time.

This let's you easily create a function to query 'myproduct'.
}
\examples{
sql <- query_db('mydb', UTC = TRUE)
sql("SELECT * FROM consumers limit 3;")


}

