% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tables.R
\name{show_tables}
\alias{show_tables}
\title{List the tables and schemas of a postgres database}
\usage{
show_tables(dbname, ...)
}
\arguments{
\item{dbname}{A valid database}

\item{...}{Other arguments passed on to methods}
}
\value{
dataframe containing the query results
}
\description{
Returns the results in a dataframe
}
\examples{
show_tables(dbname = 'mydb')

}

