% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{supported_drivers}
\alias{supported_drivers}
\title{List supported drivers}
\usage{
supported_drivers()
}
\value{
returns a vector of supported database drivers (character).
Names relate to the R packages containing the relevant driver function
}
\description{
List supported drivers
}
