% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_database_already_in_yaml}
\alias{utils_database_already_in_yaml}
\title{Database Utils}
\usage{
utils_database_already_in_yaml(dbname, file)
}
\arguments{
\item{dbname}{database name (character)}

\item{file}{file (string)}
}
\value{
true if yaml entry for dbname found, otherwise false (logical)
}
\description{
Check if a yaml contains an entry describing a given database
}
