% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_database_get_or_set_config}
\alias{utils_database_get_or_set_config}
\title{Interactive getting/setting database configurations}
\usage{
utils_database_get_or_set_config(dbname, file)
}
\arguments{
\item{dbname}{nane of database (string)}

\item{file}{yaml file to store database configuration information}
}
\value{
list describing database configuration
}
\description{
Interactive getting/setting database configurations
}
