% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_database_get_or_set_creds}
\alias{utils_database_get_or_set_creds}
\title{Create/Retrieve database credentials}
\usage{
utils_database_get_or_set_creds(dbname)
}
\arguments{
\item{dbname}{Name of database to store/retrieve credentials for (string)}
}
\value{
Invisibly returns list of username and password
}
\description{
Retrieves credentials (username and password) from os credential store using keyring package.
}
\examples{
\dontrun{
creds <- easydb::util_get_database_creds(
  service = "R-keyring-test-service",
  username = "donaldduck"
)
creds$username
creds$password
}
}
