% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_database_read_yaml}
\alias{utils_database_read_yaml}
\title{Read the database config}
\usage{
utils_database_read_yaml(dbname, file)
}
\arguments{
\item{dbname}{name of database whose config you want to read}

\item{file}{path to config yaml}
}
\value{
list with config
}
\description{
Read the database config
}
