% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_database_remove_entry_from_yaml}
\alias{utils_database_remove_entry_from_yaml}
\title{Delete database config entries from yaml}
\usage{
utils_database_remove_entry_from_yaml(dbnames, file)
}
\arguments{
\item{dbnames}{names of databases to delete from yaml (character)}

\item{file}{path to config yaml (string)}
}
\value{
Run for its side effects
}
\description{
Delete database config entries from yaml
}
