% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_database_write_yaml}
\alias{utils_database_write_yaml}
\title{Write database yaml}
\usage{
utils_database_write_yaml(
  dbname,
  driver,
  creds_required = FALSE,
  port = NULL,
  host = NULL,
  ssl_cert = NULL,
  ssl_key = NULL,
  ssl_ca = NULL,
  file,
  append = TRUE
)
}
\arguments{
\item{dbname}{name of database}

\item{driver}{name of driver}

\item{creds_required}{are credentials (username/password) required for this database (flag)}

\item{port}{database port}

\item{host}{database host}

\item{ssl_cert}{path to ssl certificate (string)}

\item{ssl_key}{path to ssl key (string)}

\item{ssl_ca}{path to ssl CA certificate (string)}

\item{file}{where config file should be located (will be produced if it doesn't already exist)}

\item{append}{should config file be appended or overwritten? Defualts to append. Don't change unless you know what you're doing}
}
\value{
path to config yaml containing the new database info (string)
}
\description{
Writes database config details into a yaml. Usernames and passwords are never saved in this yaml file
}
