% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{utils_user_input_retrieve_free_text}
\alias{utils_user_input_retrieve_free_text}
\title{get user input}
\usage{
utils_user_input_retrieve_free_text(
  message,
  default = NULL,
  type = c("string", "number")
)
}
\arguments{
\item{message}{message - tell user what information to input}

\item{default}{if user enters no data, this value will be returned (NULL)}

\item{type}{return type}
}
\value{
user input formatted according to type. If user does not enter anything will return value of \code{default}
}
\description{
get user input
}
