% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_build}
\alias{db_build}
\title{Build an easydb}
\usage{
db_build(cnf)
}
\arguments{
\item{cnf}{Path to an easydb configuration file, or the result of
\link{db_config}.}
}
\description{
Builds an SQLite database given an easydb configuration file.
}
