% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_config}
\alias{db_config}
\title{Read an easydb configuration file}
\usage{
db_config(cnf)
}
\arguments{
\item{cnf}{Path to an easydb configuration file, or the result of
\link{db_config}.}
}
\description{
Reads an easydb configuration file and returns a 'dbcnf' object.
}

