% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_dump}
\alias{db_dump}
\title{Dump an easydb database to CSVs}
\usage{
db_dump(cnf, dir)
}
\arguments{
\item{cnf}{Path to an easydb configuration file, or the result of
\link{db_config}.}

\item{dir}{Directory to write CSVs.}
}
\description{
Dumps each table in an SQLite database to CSVs
}

