% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easydb.R
\name{db_update}
\alias{db_update}
\title{Run database update expressions}
\usage{
db_update(cnf)
}
\arguments{
\item{cnf}{Path to an EasyDB configuration file, or the result of
\link{db_config}.}
}
\description{
Updates plain text source tables by evaluating R expressions in an easydb
configuration file.
}

