\name{easydb-package}
\alias{easydb-package}
\docType{package}
\title{Simple database interface for SQLite, MySQL and MS Access.}
\description{Simple database interface for SQLite, MySQL, MS Access and MS Excel. easydb is build on top of RSQLite and RODBC. It provides functions for reading, writing, listing and subsetting tables in SQLite, MySQL and MS Access. easydb allow to retrieve, write and subset database's tables in a similar way as data.frame's, using single square brackets db[,,,]. easydb internalise many database operations (otherwise tedious) such as rows and columns subsetting (no SQL needed) and connection and disconnection to / from databases.}
\details{
\tabular{ll}{Package: \tab easydb\cr
Version: \tab 0.6.5\cr
Date: \tab 2012-04-24\cr
Title: \tab Simple database interface for SQLite, MySQL and MS Access.\cr
Author: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Maintainer: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Depends: \tab R (>= 2.10.0)\cr
Suggests: \tab RODBC, RSQLite\cr
License: \tab AGPL-3\cr
URL: \tab http://easydb.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <Julien.Moeys@mark.slu.se>}

\keyword{ package }


