% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDB.R
\name{easydb_available_databases}
\alias{easydb_available_databases}
\title{Available Databases}
\usage{
easydb_available_databases(config_file)
}
\arguments{
\item{config_file}{path to yaml file containing configuration information about databases (port, host, etc. ) (string)}
}
\value{
database names (character vector)
}
\description{
List databases with configuration details stored in a configuration file.
}
\examples{
path_to_config = tempfile()
easydb_available_databases(path_to_config)
}
